pyproject.toml
[tool.poetry]
name = "jira-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for Jira Cloud integration"
authors = ["John Hull"]
readme = "README.md"
keywords = ["jira", "atlassian", "mcp", "model-context-protocol", "ai-assistant", "issue-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Bug Tracking",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
]

[tool.poetry.dependencies]
python = "^3.12"
mcp = "^1.0.0"
anyio = "^4.5"
python-dotenv = "^1.0.0"
jira = "^3.6.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
flake8 = "^6.0.0"
bandit = "^1.7.0"
pip-audit = "^2.6.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
jira-mcp = "jira_mcp.__main__:run_main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=jira_mcp",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v",
]

[tool.pylint.main]
jobs = 1

[tool.pylint.messages_control]
disable = [
    "too-few-public-methods",
    "no-member",
    "broad-exception-caught",
    "import-outside-toplevel",
    "too-many-return-statements",
    "import-error",
]

[tool.pylint.format]
max-line-length = 160
max-module-lines = 3000

[tool.pylint.design]
max-args = 10
max-locals = 25
max-returns = 6

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
